Add selection methods and web interface
Added file index.html and http backend to make user interface to get data from the DB. Created a method, which take data from the DB and add this to cache where user ask it. Method to grab data from the DB when server starting on work. Rewrote function, which generate json to make them more random to read. Error has occurred with a file go.sum, but I correct it.